Bug fixing and email content update

+++ wp-content/plugins/wc-class-booking/includes/frontend/views/class-wccb-frontend-myaccount-view.php
@@ -942,7 +942,7 @@ class WCCB_Frontend_Myaccount_View {
 								</a>
 							</td>
 							<td>
-								<?php echo (int)$value['purchased_hours'];?>
+								<?php echo (float)$value['purchased_hours'];?>
 							</td>
 							<td>
 								<?php echo (float)$value['used_hours'];?>
